Split FastPD expectation into new file
